Make self-signed cert API
